add login and register

+++ Controllers/UserController.cs
@@ -9,6 +9,8 @@ using Microsoft.AspNetCore.Hosting;
 using System.IO;
 using System.Threading.Tasks;
 using SimpleSocialNetwork.ViewModels;
+using Microsoft.AspNetCore.Identity;
+using System.Security.Claims;
 
 namespace SimpleSocialNetwork.Controllers
 {
@@ -18,11 +20,15 @@ namespace SimpleSocialNetwork.Controllers
         IHostingEnvironment _environment;
         User _user;
 
-        public UserController(IUsersRepository repository, IHostingEnvironment environment)
+        public UserController(IUsersRepository repository,
+                              IHostingEnvironment environment,
+                              IHttpContextAccessor httpContextAccessor,
+                              UserManager<User> userManager)
         {
             _repository = repository;
             _environment = environment;
-            _user = ((List<User>)_repository.Users)[0];
+            var id = userManager.GetUserId(httpContextAccessor.HttpContext.User);
+            _user = _repository.GetUserById(id);
         }
 
         public ViewResult Index()
@@ -31,9 +37,9 @@ namespace SimpleSocialNetwork.Controllers
             return View(_user);
         }
 
-        public ActionResult MainPage(int userId)
+        public ActionResult MainPage(string userId)
         {
-            if(userId == _user.UserId)
+            if(userId == _user.Id)
                 return Redirect("~/User");
             User user = _repository.GetUserById(userId);
             _repository.GetUsersMainPageInfo(user);
@@ -41,9 +47,9 @@ namespace SimpleSocialNetwork.Controllers
         }
 
 
-        public ViewResult Friends(int userId)
+        public ViewResult Friends(string userId = null)
         {
-            User user = userId == 0 ? _user:_repository.GetUserById(userId);
+            User user = userId == null ? _user:_repository.GetUserById(userId);
 
             var friends =_repository.GetUsersFriends(user);
             foreach(var friend in friends)
@@ -88,14 +94,15 @@ namespace SimpleSocialNetwork.Controllers
         [NonAction]
         public async Task UpdateAvatar(IFormFile avatar)
         {
-            string path = "/usersPhotos/avatar" + _user.UserId + ".jpg";
+            string path = "/usersPhotos/avatar" + _user.Id + ".jpg";
             using (var fileStream = new FileStream(_environment.WebRootPath + path, FileMode.Create))
             {
                 await avatar.CopyToAsync(fileStream);
             }
 
             var prevPhoto = _repository.GetUsersMainPhoto(_user);
-            _repository.Remove(prevPhoto);
+            if (prevPhoto != null)
+                _repository.Remove(prevPhoto);
 
             Post mainPhoto = new Post { Owner = _user, Date = DateTime.Now, Type = PostType.MainPhoto };
             Photo photo = new Photo { Image = path, Post = mainPhoto };
@@ -108,8 +115,7 @@ namespace SimpleSocialNetwork.Controllers
 
         public override void OnActionExecuted(ActionExecutedContext context)
         {
-            ViewBag.User = _user;
+            ViewBag.LoggedUser = _user;
         }
-
     }
 }
